Feat: Resolving Bug

+++ src/main/java/com/agroconnect/frogger/factory/DefaultDeliveryFactory.java
@@ -4,14 +4,16 @@ import com.agroconnect.frogger.entity.Delivery;
 import com.agroconnect.frogger.entity.DeliveryStatus;
 import org.springframework.stereotype.Component;
 
+import java.math.BigInteger;
+
 @Component
 public class DefaultDeliveryFactory implements DeliveryFactory {
 
     @Override
-    public Delivery createDelivery(Integer orderId, Integer deliverymanId, String thirdPartyService, String status) {
+    public Delivery createDelivery(BigInteger orderId, BigInteger deliverymanId, String thirdPartyService, String status) {
         Delivery delivery = new Delivery();
-        delivery.setOrderId(Long.valueOf(orderId));
-        delivery.setDeliverymanId(Long.valueOf(deliverymanId));
+        delivery.setOrderId(orderId);
+        delivery.setDeliverymanId(deliverymanId);
         delivery.setThirdPartyService(thirdPartyService);
         delivery.setDeliveryStatus(DeliveryStatus.valueOf(status));
         return delivery;
