+++ src/main/java/com/agroconnect/frogger/factory/DeliveryFactory.java
@@ -2,6 +2,8 @@ package com.agroconnect.frogger.factory;
 
 import com.agroconnect.frogger.entity.Delivery;
 
+import java.math.BigInteger;
+
 public interface DeliveryFactory {
-    Delivery createDelivery(Integer orderId, Integer deliverymanId, String thirdPartyService, String status);
+    Delivery createDelivery(BigInteger orderId, BigInteger deliverymanId, String thirdPartyService, String status);
 }
